Fallback to SingingCertificate reference when processing KeyInfo to identify the signer's certificate.
Supports signatures without KeyInfo when the SigningCertificate property contains a single reference (most common use case)
